Backend/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy 
jwt 
python-dotenv
passlib
bcrypt==4.0.1 
python-multipart
python-jose[cryptography]
pypdf
langchain
langchain_chroma
langchain_community
langchain_huggingface
langchain_ollama
streamlit
python-docx
PyMuPDF
faiss-cpu
rank-bm25
sentence-transformers
fastapi[all]
nltk
langchain_groq
pytest
pytest-asyncio